+++ src/app/effects/user.effects.ts
@@ -3,6 +3,7 @@ import { Effect, Actions } from '@ngrx/effects';
 import { ActionTypes, UserActions } from '../actions/user.actions';
 import { Action } from '@ngrx/store';
 import { ApiService } from '../services/api.service';
+import { ResponseParseService } from '../services/response-parse.service';
 import { User } from '../models';
 
 @Injectable()
@@ -11,12 +12,14 @@ export class UserEffects {
   constructor(
     private actions$: Actions,
     private userActions: UserActions,
-    private api: ApiService){ }
+    private api: ApiService,
+    private responseParser: ResponseParseService
+    ){ }
 
     @Effect() loadUserProfile$ = this.actions$
       .ofType(ActionTypes.LOAD_PROFILE)
       .switchMap(() => this.api.getUserDetail())
-      .map((response:any) => this.api.createUserObj(response))
+      .map((response:any) => this.responseParser.createUserObj(response))
       .map((user: User) => this.userActions.loadProfileSuccess(user))
