+++ src/app/models/tweet.ts
@@ -1,5 +1,8 @@
 import { Base } from './base';
+import { User } from './user';
 
 export class Tweet extends Base {
-  text: string;
+  created_at: string
+  text:       string;
+  user:       User
 }
