+++ src/app/models/urls.ts
@@ -1,0 +1,7 @@
+export class Urls {
+  constructor(
+    url:          string,
+    expanded_url: string,
+    display_url:  string,
+    indices:      Array<number>) {}
+}
