+++ src/app/models/user.ts
@@ -1,7 +1,10 @@
 import { UserProfile } from './user-profile';
 import { Base } from './base';
+import { Entities } from './entities';
+
 export class User extends Base {
   name:              string;
   screen_name:       string;
   profile_image_url: string;
+  // entities:          Entities;
 }
